+++ src/util/util_agente.py
@@ -1,6 +1,5 @@
 # Utilitario para crear y ejecutar agentes
-from langchain_core.prompts import ChatPromptTemplate
-from langgraph.prebuilt import create_react_agent
+from langchain.agents import create_agent
 
 # Utilitario para el modelo de lenguaje
 from langchain_google_genai import ChatGoogleGenerativeAI
@@ -8,26 +7,49 @@ from langchain_google_genai import ChatGoogleGenerativeAI
 # Manejo de memoria del agente
 from langgraph.checkpoint.memory import InMemorySaver
 
-# Manejo de prompts
-from langchain_core.prompts import ChatPromptTemplate
+import json
 
 def crearAgente(
-    llm: ChatGoogleGenerativeAI, contexto: ChatPromptTemplate, tools: list | None = None, memoria=None
+    llm: ChatGoogleGenerativeAI, contexto: str, tools: list | None = None, memoria=None
 ):
     if tools is None:
         tools = []
     if memoria is None:
         memoria = InMemorySaver()
-    agente = create_react_agent(
-        model=llm, tools=tools, checkpointer=memoria, prompt=contexto,
+    agente = create_agent(
+        model=llm, tools=tools, checkpointer=memoria, system_prompt=contexto,
+    )
+    return agente
+
+def crearAgenteSinMemoria(
+    llm: ChatGoogleGenerativeAI, contexto: str, tools: list | None
+):
+    if tools is None:
+        tools = []
+    agente = create_agent(
+        model=llm, tools=tools, system_prompt=contexto,
     )
     return agente
 
 def ejecutar(agente, consulta: str = "", config=None, verbose: bool = True):
+    payload = {"messages": [{"role": "user", "content": consulta}]}
+    
+    respuesta = agente.invoke(payload, config=config)
     try:
-        respuesta = agente.invoke({"messages": [{"role": "user", "content": consulta}]}, config=config)
         if not verbose:
             return respuesta
         return respuesta["messages"][-1].content
+    except Exception as e:
+        raise Exception(f'Error en la ejecución del agente: {e}')
+
+def ejecutarSinMemoria(agente, consulta: str = "", verbose: bool = True):
+    payload = {"messages": [{"role": "user", "content": consulta}]}
+    
+    respuesta = agente.invoke(payload)
+    try:
+        if not verbose:
+            return respuesta
+        respuesta = respuesta["messages"][-1].content.replace("```json", "").replace("```", "")
+        return json.loads(respuesta)
     except Exception as e:
         raise Exception(f'Error en la ejecución del agente: {e}')