+++ src/util/util_llm.py
@@ -4,7 +4,7 @@ from src.util.util_credenciales import obtenerAPI
 
 def obtenerModelo() -> ChatGoogleGenerativeAI:
     return ChatGoogleGenerativeAI(
-        model="gemini-1.5-flash",
+        model="gemini-2.5-flash",
         api_key=obtenerAPI("CONF-GOOGLE-API-KEY"),
     )