+++ CHANGELOG.md
@@ -1,3 +1,59 @@
+## [1.8.3](https://github.com/NickSalA/LuminAIBackend/compare/v1.8.2...v1.8.3) (2025-11-07)
+
+
+### Bug Fixes
+
+* Actualización de las respuestas y formateo JSON para la calificacion para mayor claridad ([c553335](https://github.com/NickSalA/LuminAIBackend/commit/c55333557c4c32f2db6411ad8966060b8c4e4bae))
+
+## [1.8.2](https://github.com/NickSalA/LuminAIBackend/compare/v1.8.1...v1.8.2) (2025-11-06)
+
+
+### Bug Fixes
+
+* Preguntas generadas en formato JSON y reestructuracion del prompt de AgentePregunta ([b961356](https://github.com/NickSalA/LuminAIBackend/commit/b96135674437e3b194bb33d2b84f6567d5855f0d))
+
+## [1.8.1](https://github.com/NickSalA/LuminAIBackend/compare/v1.8.0...v1.8.1) (2025-11-06)
+
+
+### Bug Fixes
+
+* Refactorizando la IA en Agente Respuestas y retroalimentación ([36a36a8](https://github.com/NickSalA/LuminAIBackend/commit/36a36a8d82ef9ea1b3b665ed39ab70e336729397))
+
+# [1.8.0](https://github.com/NickSalA/LuminAIBackend/compare/v1.7.1...v1.8.0) (2025-11-06)
+
+
+### Features
+
+* Implementar creación de agentes sin memoria y mejorar la gestión de sesiones en la API ([d24f59f](https://github.com/NickSalA/LuminAIBackend/commit/d24f59f3b7a24c544696bd00708094a1424f579b))
+
+## [1.7.1](https://github.com/NickSalA/LuminAIBackend/compare/v1.7.0...v1.7.1) (2025-11-06)
+
+
+### Bug Fixes
+
+* Corrección de la forma en que se llamaba el config de la respuesta del agente ([2acc0c1](https://github.com/NickSalA/LuminAIBackend/commit/2acc0c1a899c3e03c0781e8eac3e6f0235d4a72e))
+
+# [1.7.0](https://github.com/NickSalA/LuminAIBackend/compare/v1.6.0...v1.7.0) (2025-11-06)
+
+
+### Features
+
+* Implementar CORS y sesión en la API, actualizar rutas y mejorar el modelo de lenguaje ([bdb7588](https://github.com/NickSalA/LuminAIBackend/commit/bdb75889142edd0bb4839519bab9130c0cbdf595))
+
+# [1.6.0](https://github.com/NickSalA/LuminAIBackend/compare/v1.5.0...v1.6.0) (2025-11-03)
+
+
+### Features
+
+* Mejora del prompt de respuesta y preguntas ([926a9d9](https://github.com/NickSalA/LuminAIBackend/commit/926a9d90b4ead598b2da9b222d3b133628402f60))
+
+# [1.5.0](https://github.com/NickSalA/LuminAIBackend/compare/v1.4.0...v1.5.0) (2025-10-31)
+
+
+### Features
+
+* Actualizar credenciales para la api de azure ([6d22d9d](https://github.com/NickSalA/LuminAIBackend/commit/6d22d9d38fec6ad65ec0a26e1bb0867a3404caf4))
+
 # [1.4.0](https://github.com/NickSalA/LuminAIBackend/compare/v1.3.0...v1.4.0) (2025-10-25)